requirements.txt
torch
numpy
Pillow
scikit-learn
kagglehub
jupyter
